splitted test for PhotoAlbumService into -Test and -IT
