+++ src/test/java/de/thg/photoalbum/services/PhotoAlbumServiceTest.java
@@ -8,18 +8,14 @@ import org.apache.commons.io.FilenameUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
@@ -30,7 +26,6 @@ import static org.assertj.core.api.Assertions.assertThat;
  *
  */
 @SpringBootTest
-@ActiveProfiles("test")
 public class PhotoAlbumServiceTest {
 
 	private static Path sourcepath1;
@@ -44,9 +39,6 @@ public class PhotoAlbumServiceTest {
 
 	@Inject
 	private ImageRepository imageRepository;
-
-	@Value("${photoalbum.prefix:jpg}")
-	private String PREFIX;
 
 	@BeforeAll
 	static void beforeAll() throws IOException {
@@ -107,54 +99,5 @@ public class PhotoAlbumServiceTest {
 		assertThat(tempDir).doesNotExist();
 	}
 
-	@Test
-	public void testCreateOrUpdatePhotoAlbum() {
-		imageRepository.deleteAll();
-		assertThat(imageRepository.findAll()).as("table image is not empty").isEmpty();
-		assertThat(params.isDebug()).isFalse();
-		assertThat(new File(params.getTarget())).isDirectory();
-//		assertThat(new File(params.getTarget()).listFiles()).isEmpty();
-		for (String source : params.getSources()) {
-			assertThat(new File(source)).isDirectory();
-			assertThat(new File(source).listFiles()).isNotEmpty();
-		}
-		List<Image> imageList = underTest.createOrUpdatePhotoAlbum(params);
-		assertThat(imageList).hasSize(2);
-		for (Image thisImage : imageList) {
-			assertThat(thisImage.getTempFile()).isNotNull();
-		}
-		imageRepository.saveAll(imageList);
-		for (Image imageFromDb : imageRepository.findAll()) {
-			assertThat(imageFromDb.getCreationDate().equals(imageList.get(0).getCreationDate()) ||
-					imageFromDb.getCreationDate().equals(imageList.get(1).getCreationDate())).as("error in DB").isTrue();
-			assertThat(imageFromDb.getTempFile()).isNull();
-			assertThat(imageFromDb.getFilename()).as("wrong filename").startsWith(PREFIX);
-		}
-	}
 
-	@Test
-	@Disabled
-	void renameImageFiles() throws IOException {
-		String sourcedir = "/home/tom/Bilder/Test";
-		assertThat(new File(sourcedir).listFiles()).isNotEmpty();
-		List<Image> images = underTest.renameImageFiles(sourcedir);
-		assertThat(images).isNotEmpty();
-		assertThat(images.stream().map(Image::getFilename).filter(s -> !s.matches("\\d{8}_\\d{6}.jpg")).toList()).isEmpty();
-		Collections.sort(images);
-		for (Image image : images) {
-			System.out.println(image.getFilename());
-		}
-
-//		System.out.println();
-//		sourcedir = "/home/tom/Bilder/PhotoAlbum";
-//		assertThat(new File(sourcedir).listFiles()).isNotEmpty();
-//		assertThat(Arrays.stream(new File(sourcedir).listFiles()).map(File::getName).filter(s -> s.matches("\\d{8}_\\d{6}.jpg")).toList()).isEmpty();
-//		images = underTest.renameImageFiles(sourcedir);
-//		assertThat(images).isNotEmpty();
-//		assertThat(images.stream().map(Image::getFilename).filter(s -> !s.matches("\\d{8}_\\d{6}.jpg")).toList()).isEmpty();
-//		Collections.sort(images);
-//		for (Image image : images) {
-//			System.out.println(image.getFilename());
-//		}
-	}
 }
