--- src/test/resources/application-tc.properties
@@ -1,2 +1,0 @@
-#spring.datasource.driver-class-name=org.testcontainers.jdbc.ContainerDatabaseDriver
-#spring.datasource.url=jdbc:tc:postgresql:15.3-alpine3.18:///photoalbum?TC_INITSCRIPT=file:src/main/resources/db/migration/V1__create_image_table.sql
